refactor(rename): rename "meetup" → "event" across plugin, package, and copy
Legal flagged "Meetup" as a trademark (Meetup.com), so all AAIF forward-facing
wording moves to the "AAIF Community Event(s)" brand. This covers the repo side
of that rename:

- Identifiers: plugin id aaif-meetups → aaif-events, Python package
  lib/aaif_meetups → lib/aaif_events (+ all imports), repo-URL refs → aaif/events,
  install commands, display name → "AAIF Community Events Toolkit".
- Forward-facing copy: all 9 skill descriptions + speaker bio/invite bodies,
  README/CONTRIBUTING prose, manifest descriptions → "event".
- Kept: CHANGELOG dated history, and generic in-sentence "an AAIF event" titles.

Live Google Drive assets (templates, forms, guideline docs, banners across all
chapters) were migrated separately. GitHub repo rename meetups → events is an
admin follow-up; the code already points at the new URL.
