+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-# Pre-commit hooks for the AAIF Meetups plugin marketplace.
+# Pre-commit hooks for the AAIF Community Events plugin marketplace.
 #   Setup:  pipx install pre-commit && pre-commit install
 #   Run:    pre-commit run --all-files
 #   Update: pre-commit autoupdate
