+++ skills/aaif-backup/SKILL.md
@@ -38,7 +38,7 @@ python3 ${CLAUDE_SKILL_DIR}/scripts/backup.py --dest /some/dir
 ```
 
 - `<dest>` defaults to **`./backups`** (relative to where you run it — inside the
-  `meetups/` repo when run from there). `backups/` is **git-ignored**, so the binary
+  `events/` repo when run from there). `backups/` is **git-ignored**, so the binary
   snapshots never enter the repo.
 - Filenames are UTC timestamps, so a folder listing is the version history in order.
 - The target's type decides the format: a Google Sheet is exported to `.xlsx`, a Doc to
