+++ skills/aaif-dayof-slides/SKILL.md
@@ -1,6 +1,6 @@
 ---
 name: aaif-dayof-slides
-description: Turn an event's tracker entry into the slide text for the AAIF "Day of Event" deck. Use when asked to fill/write the day-of slides or event deck for an AAIF meetup.
+description: Turn an event's tracker entry into the slide text for the AAIF "Day of Event" deck. Use when asked to fill/write the day-of slides or event deck for an AAIF event.
 argument-hint: '[event title / paste tracker entry]'
 ---
