+++ skills/aaif-event-status/scripts/event_status.py
@@ -9,7 +9,7 @@ import pathlib
 import sys
 
 sys.path.insert(0, str(pathlib.Path(__file__).resolve().parents[3] / "lib"))
-from aaif_meetups import office, tracker  # noqa: E402
+from aaif_events import office, tracker  # noqa: E402
 
 DUE_SOON_DAYS = 7
