+++ skills/aaif-event-status/scripts/luma_stats.py
@@ -11,7 +11,7 @@ import pathlib
 import sys
 
 sys.path.insert(0, str(pathlib.Path(__file__).resolve().parents[3] / "lib"))
-from aaif_meetups import luma, office, tracker  # noqa: E402
+from aaif_events import luma, office, tracker  # noqa: E402
 
 COUNT_KEYS = ("approved", "pending_approval", "waitlist", "invited", "declined", "checked_in")
